Hint metadata and global units can be combined
This never occurred to me after months of wondering about how I might schedule
units to take advantage of AWS auto-scaling. It's obvious in retrospect, but
there's a mental speed-bump: how can a unit be "global" if it runs only on some
subset, according to metadata? This hint should be sufficient to avoid that
mental block.