os/kernel-modules: fix mounting of overlay
I noticed that as of 1353.4.0 (maybe earlier) the lib-modules.mount
unit was no longer started on boot. Checking `systemctl status
local-fs.target` returned:

```
local-fs.target: Wants dependency dropin
/etc/systemd/system/local-fs.target.wants/lib-modules.mount is not a
symlink, ignoring.
```

To fix this, we can create the file in /etc/systemd/system and then
enable it to create the symlink. I tested this on my machines and found
it to work.