[WIP] os/quickstart: remove fleet, reword suitably
This is a first pass at a new structure for this document.

I haven't filled in most of the words, I mostly just want any ideas
around whether this structure/format makes sense, what headings / topics
should be covered, etc.

The thing I'm least sure about is whether the idea of "three tools" is
the right way to present this, and whether service discovery vs
provisioning matters as a member of that list.