os/booting-on-ec2: update to CLC 2.1 syntax
Switch to always creating a filesystem on the ephemeral disk.
SSD-backed instance stores are not preformatted, and the others are
formatted with ext3. We shouldn't universally recommend ext3 -- ext4
is better -- and we shouldn't (and, with wipe_filesystem, can't)
conditionally produce an ext3 or ext4 ephemeral volume depending on the
instance type. Thus, always format as ext4; mkfs.ext4 is reasonably
fast anyway.